services/onnx-asr/pyproject.toml
[project]
name = "onnx-asr"
version = "1.0.0"
description = "ONNX-based ASR service using nemo-parakeet-tdt-0.6b-v2"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "aiokafka>=0.11.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
    "onnxruntime-gpu>=1.17.0",
    "huggingface-hub>=0.20.0",
    "tokenizers>=0.15.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_backend"

[tool.setuptools]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
